@@ -28,8 +28,6 @@ public class User {
     @Column(nullable = false, unique = true, length = 20)
     private String username;
 
-    @Column(nullable = false, length = 60) // for hashed passwords
-    private String password;
 
     @Column(nullable = false)
     private LocalDate dateCreated = LocalDate.now(); // column name now matches Java field
@@ -37,11 +35,11 @@ public class User {
     // Constructors
     public User() {}
 
-    public User(String username, String password) {
+    public User(String username) {
         this.username = username;
-        this.password = password;
         this.dateCreated = LocalDate.now();
     }
+
 
     // Getters and setters
     public Integer getId() { return id; }
@@ -49,9 +47,6 @@ public class User {
 
     public String getUsername() { return username; }
     public void setUsername(String username) { this.username = username; }
-
-    public String getPassword() { return password; }
-    public void setPassword(String password) { this.password = password; }
 
     public LocalDate getDateCreated() { return dateCreated; }
     public void setDateCreated(LocalDate dateCreated) { this.dateCreated = dateCreated; }
