@@ -5,29 +5,22 @@ import org.flashcard.application.dto.DeckDTO;
 import org.flashcard.controllers.DeckController;
 import org.flashcard.controllers.UserController;
 
+import org.flashcard.controllers.observer.Observer; // NEW
+
 import javax.swing.*;
 import java.awt.*;
 import java.util.List;
 
-/**
- * EditDeckView - vy för att redigera ett valt deck:
- * - visa lista med flashcards (med delete-knappar)
- * - lägga till flashcards
- * - radera hela leken
- * - back/return
- */
-public class EditDeckView extends JPanel {
+public class EditDeckView extends JPanel implements Observer<List<FlashcardDTO>> {  // NEW
 
     private final DeckController deckController;
     private final UserController userController;
     private final AppFrame appFrame;
 
-    // State
     private DeckDTO currentDeck;
 
-    // UI
     private JLabel headerLabel;
-    private JPanel cardsListPanel; // innehåller rader för varje kort
+    private JPanel cardsListPanel;
     private JScrollPane cardsScroll;
     private JTextField newFrontField;
     private JTextField newBackField;
@@ -41,26 +34,30 @@ public class EditDeckView extends JPanel {
         this.userController = userController;
         this.appFrame = appFrame;
 
+        // registrera observer
+        deckController.getFlashcardsObservable().addListener(this);
+
         setLayout(new BorderLayout());
         setBackground(new Color(245, 245, 245));
         initComponents();
     }
 
     private void initComponents() {
-        // Header
+
+
         JPanel top = new JPanel(new BorderLayout());
         top.setBackground(new Color(245, 245, 245));
         top.setBorder(BorderFactory.createEmptyBorder(12, 16, 12, 16));
 
-        backButton = new JButton("← Tillbaka");
+        backButton = new JButton("← Back");
         backButton.addActionListener(e -> appFrame.navigateTo("MyDecks"));
         top.add(backButton, BorderLayout.WEST);
 
-        headerLabel = new JLabel("Redigera lek", SwingConstants.CENTER);
+        headerLabel = new JLabel("Edit Deck", SwingConstants.CENTER);
         headerLabel.setFont(new Font("SansSerif", Font.BOLD, 20));
         top.add(headerLabel, BorderLayout.CENTER);
 
-        deleteDeckButton = new JButton("Radera lek");
+        deleteDeckButton = new JButton("Delete Deck");
         deleteDeckButton.setBackground(new Color(220, 60, 60));
         deleteDeckButton.setForeground(Color.WHITE);
         deleteDeckButton.addActionListener(e -> confirmAndDeleteDeck());
@@ -68,7 +65,6 @@ public class EditDeckView extends JPanel {
 
         add(top, BorderLayout.NORTH);
 
-        // Center: cards list
         cardsListPanel = new JPanel();
         cardsListPanel.setLayout(new BoxLayout(cardsListPanel, BoxLayout.Y_AXIS));
         cardsListPanel.setBackground(Color.WHITE);
@@ -79,7 +75,6 @@ public class EditDeckView extends JPanel {
         cardsScroll.getVerticalScrollBar().setUnitIncrement(14);
         add(cardsScroll, BorderLayout.CENTER);
 
-        // Bottom: add card form + status
         JPanel bottom = new JPanel();
         bottom.setLayout(new BoxLayout(bottom, BoxLayout.Y_AXIS));
         bottom.setBackground(new Color(245, 245, 245));
@@ -92,18 +87,18 @@ public class EditDeckView extends JPanel {
         gbc.fill = GridBagConstraints.HORIZONTAL;
 
         gbc.gridx = 0; gbc.gridy = 0; gbc.weightx = 0;
-        form.add(new JLabel("Framsida:"), gbc);
+        form.add(new JLabel("Front Side:"), gbc);
         newFrontField = new JTextField();
         gbc.gridx = 1; gbc.gridy = 0; gbc.weightx = 1;
         form.add(newFrontField, gbc);
 
         gbc.gridx = 0; gbc.gridy = 1; gbc.weightx = 0;
-        form.add(new JLabel("Baksida:"), gbc);
+        form.add(new JLabel("Back Side:"), gbc);
         newBackField = new JTextField();
         gbc.gridx = 1; gbc.gridy = 1; gbc.weightx = 1;
         form.add(newBackField, gbc);
 
-        addCardButton = new JButton("+ Lägg till kort");
+        addCardButton = new JButton("+ Add Card");
         addCardButton.setPreferredSize(new Dimension(150, 36));
         addCardButton.setBackground(new Color(60, 160, 80));
         addCardButton.setForeground(Color.WHITE);
@@ -116,7 +111,6 @@ public class EditDeckView extends JPanel {
         bottom.add(form);
         bottom.add(formRow);
 
-        // status
         statusLabel = new JLabel(" ");
         statusLabel.setFont(new Font("SansSerif", Font.PLAIN, 12));
         statusLabel.setForeground(new Color(0, 120, 0));
@@ -126,17 +120,14 @@ public class EditDeckView extends JPanel {
         add(bottom, BorderLayout.SOUTH);
     }
 
-    /**
-     * Ladda deck i edit-läget (anropa från AppFrame eller MyDecksView när användaren klickar "Edit")
-     */
     public void loadDeck(int deckId) {
         try {
             currentDeck = deckController.getDeckById(deckId);
-            headerLabel.setText("Redigera: " + currentDeck.getTitle());
-            statusLabel.setText("Laddat lek (ID " + currentDeck.getId() + ")");
+            headerLabel.setText("Edit: " + currentDeck.getTitle());
+            statusLabel.setText("Loaded deck (ID " + currentDeck.getId() + ")");
             refreshCardsList();
         } catch (Exception e) {
-            JOptionPane.showMessageDialog(this, "Kunde inte ladda lek: " + e.getMessage());
+            JOptionPane.showMessageDialog(this, "Couldn't load the Deck: " + e.getMessage());
         }
     }
 
@@ -149,11 +140,10 @@ public class EditDeckView extends JPanel {
             return;
         }
 
-        // Hämta kort via controllern (säkrare än att förlita sig på lazy-listor)
         List<FlashcardDTO> cards = deckController.getFlashcardsForDeck(currentDeck.getId());
 
         if (cards.isEmpty()) {
-            JLabel empty = new JLabel("Inga kort i denna lek. Lägg till ett nedan.");
+            JLabel empty = new JLabel("No cards in this deck. Add one below.");
             empty.setBorder(BorderFactory.createEmptyBorder(10,10,10,10));
             cardsListPanel.add(empty);
         } else {
@@ -171,7 +161,6 @@ public class EditDeckView extends JPanel {
         row.setBackground(Color.WHITE);
         row.setBorder(BorderFactory.createMatteBorder(0,0,1,0, new Color(230,230,230)));
 
-        // Left: front/back text
         JPanel text = new JPanel();
         text.setLayout(new BoxLayout(text, BoxLayout.Y_AXIS));
         text.setBackground(Color.WHITE);
@@ -182,11 +171,10 @@ public class EditDeckView extends JPanel {
         text.add(front);
         text.add(back);
 
-        // Right: action buttons
         JPanel actions = new JPanel(new FlowLayout(FlowLayout.RIGHT, 6, 6));
         actions.setBackground(Color.WHITE);
 
-        JButton deleteBtn = new JButton("Radera");
+        JButton deleteBtn = new JButton("Delete");
         deleteBtn.setBackground(new Color(220,60,60));
         deleteBtn.setForeground(Color.WHITE);
         deleteBtn.addActionListener(e -> confirmAndDeleteCard(card.getId()));
@@ -203,47 +191,50 @@ public class EditDeckView extends JPanel {
 
     private void handleAddCard() {
         if (currentDeck == null) {
-            JOptionPane.showMessageDialog(this, "Ingen lek vald.");
+            JOptionPane.showMessageDialog(this, "No Deck chosen.");
             return;
         }
         String front = newFrontField.getText().trim();
         String back = newBackField.getText().trim();
 
         if (front.isBlank() || back.isBlank()) {
-            JOptionPane.showMessageDialog(this, "Både framsida och baksida krävs.");
+            JOptionPane.showMessageDialog(this, "Both front and back are required.");
             return;
         }
 
         try {
             deckController.addFlashcard(currentDeck.getId(), front, back);
-            // rensa formulär och uppdatera listan
             newFrontField.setText("");
             newBackField.setText("");
-            statusLabel.setText("Kort tillagt!");
-            // uppdatera currentDeck DTO (så t.ex. cardCount blir korrekt)
+            statusLabel.setText("Card added.");
+
             currentDeck = deckController.getDeckById(currentDeck.getId());
             refreshCardsList();
+
         } catch (Exception e) {
-            JOptionPane.showMessageDialog(this, "Kunde inte lägga till kort: " + e.getMessage());
+            JOptionPane.showMessageDialog(this, "Couldn't add the Card " + e.getMessage());
         }
     }
 
     private void confirmAndDeleteCard(Integer cardId) {
-        int res = JOptionPane.showConfirmDialog(this,
-                "Är du säker på att du vill ta bort detta kort?",
-                "Bekräfta radering",
+        int res = JOptionPane.showConfirmDialog(
+                this,
+                "Are you sure you want to delete this card?",
+                "Confirm deletion",
                 JOptionPane.YES_NO_OPTION,
-                JOptionPane.WARNING_MESSAGE);
+                JOptionPane.WARNING_MESSAGE
+        );
+
         if (res == JOptionPane.YES_OPTION) {
             try {
                 deckController.deleteFlashcard(cardId);
-                statusLabel.setText("Kort raderat.");
-                if (currentDeck != null) {
-                    currentDeck = deckController.getDeckById(currentDeck.getId());
-                }
+                statusLabel.setText("Card deleted.");
+
+                currentDeck = deckController.getDeckById(currentDeck.getId());
                 refreshCardsList();
+
             } catch (Exception e) {
-                JOptionPane.showMessageDialog(this, "Kunde inte radera kort: " + e.getMessage());
+                JOptionPane.showMessageDialog(this, "Couldn't delete card " + e.getMessage());
             }
         }
     }
@@ -251,20 +242,31 @@ public class EditDeckView extends JPanel {
     private void confirmAndDeleteDeck() {
         if (currentDeck == null) return;
 
-        int res = JOptionPane.showConfirmDialog(this,
-                "Är du säker? Detta tar bort leken och alla kort i den.",
-                "Bekräfta radering",
+        int res = JOptionPane.showConfirmDialog(
+                this,
+                "Are you sure? This will delete the deck and all cards in it.",
+                "Confirm deletion",
                 JOptionPane.YES_NO_OPTION,
-                JOptionPane.WARNING_MESSAGE);
+                JOptionPane.WARNING_MESSAGE
+        );
+
         if (res == JOptionPane.YES_OPTION) {
             try {
                 deckController.deleteDeck(currentDeck.getId());
-                JOptionPane.showMessageDialog(this, "Leken raderades.");
-                // gå tillbaka till MyDecks och uppdatera
+                JOptionPane.showMessageDialog(this, "Deck deleted.");
                 appFrame.navigateTo("MyDecks");
+
             } catch (Exception e) {
-                JOptionPane.showMessageDialog(this, "Kunde inte radera lek: " + e.getMessage());
+                JOptionPane.showMessageDialog(this, "Couldn't delete Deck " + e.getMessage());
             }
         }
+    }
+
+
+    // OBSERVER CALLBACK METHOD
+    @Override
+    public void notify(List<FlashcardDTO> updatedCards) {
+        // Ladda om kortlistan automatiskt
+        SwingUtilities.invokeLater(this::refreshCardsList);
     }
 }
