@@ -3,12 +3,13 @@ package org.flashcard.testview;
 import org.flashcard.application.dto.FlashcardDTO;
 import org.flashcard.controllers.DeckController;
 import org.flashcard.controllers.StudyController;
+import org.flashcard.controllers.observer.Observer;  // NEW
 import org.flashcard.models.dataclasses.Deck;
 
 import javax.swing.*;
 import java.awt.*;
 
-public class StudyViewTest extends JPanel {
+public class StudyView extends JPanel implements Observer<FlashcardDTO> {
 
     private final StudyController studyController;
     private final AppFrame appFrame;
@@ -18,16 +19,29 @@ public class StudyViewTest extends JPanel {
     private JPanel controlsPanel;
     private JButton showAnswerButton;
     private JPanel ratingPanel;
+    private JButton nextButton;
     private JPanel intervalPanel;
     private JButton nextButton; // NY: För Study All mode
 
     private FlashcardDTO currentCard;
     private String currentStrategy; // "today" eller "all"
 
-    public StudyViewTest(DeckController deckController, StudyController studyController, AppFrame appFrame) {
+    // Observer for session finished
+    private final Observer<Boolean> finishedListener = finished -> {
+        if (finished != null && finished) {
+            handleSessionFinished();
+        }
+    };
+
+    public StudyView(StudyController studyController, DeckController deckController, AppFrame appFrame) {
         this.deckController = deckController;
         this.studyController = studyController;
         this.appFrame = appFrame;
+
+        // Register observers
+        studyController.getCurrentCardObservable().addListener(this);
+        studyController.getSessionFinishedObservable().addListener(finishedListener);
+
         setLayout(new BorderLayout());
         initComponents();
     }
@@ -51,7 +65,7 @@ public class StudyViewTest extends JPanel {
         controlsPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 20, 20));
         controlsPanel.setPreferredSize(new Dimension(800, 100));
 
-        showAnswerButton = new JButton("Visa Svar");
+        showAnswerButton = new JButton("Show Answer");
         showAnswerButton.setFont(new Font("SansSerif", Font.BOLD, 16));
         showAnswerButton.setPreferredSize(new Dimension(200, 50));
         showAnswerButton.addActionListener(e -> showBack());
@@ -63,6 +77,7 @@ public class StudyViewTest extends JPanel {
         addRatingIntervals();
         ratingPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 10, 0));
         ratingPanel.setVisible(false);
+
         createRatingButton("Again", "again", new Color(255, 80, 80));
         createRatingButton("Hard", "hard", new Color(255, 165, 0));
         createRatingButton("Medium", "medium", new Color(70, 130, 180));
@@ -70,13 +85,12 @@ public class StudyViewTest extends JPanel {
         ratingWrapper.add(intervalPanel);
         ratingWrapper.add(ratingPanel);
 
-        // Next Button (För 'all' mode)
-        nextButton = new JButton("Nästa Kort ->");
+        nextButton = new JButton("Next Card ->");
         nextButton.setPreferredSize(new Dimension(200, 50));
         nextButton.setBackground(new Color(60, 120, 240));
         nextButton.setForeground(Color.WHITE);
         nextButton.setFont(new Font("SansSerif", Font.BOLD, 16));
-        nextButton.addActionListener(e -> loadNextCard()); // Bara ladda nästa, ingen rating
+        nextButton.addActionListener(e -> studyController.nextCard());
         nextButton.setVisible(false);
 
         controlsPanel.add(showAnswerButton);
@@ -98,23 +112,29 @@ public class StudyViewTest extends JPanel {
 
     public void initSession(String strategy) {
         this.currentStrategy = strategy;
-        loadNextCard();
+
+        // Previously you called loadNextCard() here, now StudyController does it
     }
 
-    private void loadNextCard() {
-        currentCard = studyController.nextCard();
+    // Observer callback for new card
+    @Override
+    public void notify(FlashcardDTO card) {
+        SwingUtilities.invokeLater(() -> showNewCard(card));
+    }
 
-        if (currentCard == null) {
-            JOptionPane.showMessageDialog(this, "Passet är slut!");
-            appFrame.navigateTo("Home"); // Gå tillbaka dit vi kom ifrån egentligen
+    private void showNewCard(FlashcardDTO card) {
+        this.currentCard = card;
+
+        if (card == null) {
+            handleSessionFinished();
             return;
         }
 
-        cardTextArea.setText(currentCard.getFront());
+        cardTextArea.setText(card.getFront());
 
         showAnswerButton.setVisible(true);
         ratingPanel.setVisible(false);
-        nextButton.setVisible(false); // Dölj alltid först
+        nextButton.setVisible(false);
 
         controlsPanel.revalidate();
         controlsPanel.repaint();
@@ -122,7 +142,10 @@ public class StudyViewTest extends JPanel {
 
     private void showBack() {
         if (currentCard != null) {
-            cardTextArea.setText(currentCard.getFront() + "\n\n----------------\n\n" + currentCard.getBack());
+            cardTextArea.setText(currentCard.getFront()
+                    + "\n\n----------------\n\n"
+                    + currentCard.getBack());
+
             showAnswerButton.setVisible(false);
 
             // Visa rätt kontroller baserat på strategi
@@ -137,9 +160,9 @@ public class StudyViewTest extends JPanel {
     private void applyRating(String rating) {
         try {
             studyController.applyRating(rating, currentCard.getId());
-            loadNextCard();
+            studyController.nextCard();  // Observer will handle UI update
         } catch (Exception e) {
-            JOptionPane.showMessageDialog(this, "Fel vid rating: " + e.getMessage());
+            JOptionPane.showMessageDialog(this, "Error while rating: " + e.getMessage());
         }
     }
     private void addRatingIntervals(){
@@ -152,5 +175,11 @@ public class StudyViewTest extends JPanel {
 //        intervalPanel.add(new JLabel(String.valueOf(deckController.showEstimatedDate("easy", currentCard.getId()))));
 
 
+    }
+
+    // Called when sessionFinishedObservable fires
+    private void handleSessionFinished() {
+        JOptionPane.showMessageDialog(this, "The session is over!");
+        appFrame.navigateTo("Home");
     }
 }
